Move test_find_unread_includes_edited_comments to test_views out of test comment model
